requirements.txt
# ckeditor
asgiref==3.5.2
cssselect2==0.6.0
Django==4.0.5
django-admin-rangefilter==0.8.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
djangorestframework==3.13.1
easy-thumbnails==2.8.1
lxml==4.9.0
mssql-django==1.1.3
mysqlclient==2.1.0
Pillow==9.1.1
psycopg2==2.9.3
pyodbc==4.0.32
pytz==2022.1
reportlab==3.6.10
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tzdata==2022.1
webencodings==0.5.1
